stac-fastapi/requirements.txt
stac-fastapi.api==2.3.0
stac-fastapi.types==2.3.0
stac-fastapi.extensions==2.3.0
stac-fastapi.pgstac==2.3.0
uvicorn==0.18.2